+++ gateway/utils.go
