+++ memorystore/main.go
